requirements.txt
Flask==2.2.5
Flask-SocketIO==5.3.3
Pillow==10.0.1
